html bakery

+++ bakehtml.py
@@ -6,9 +6,21 @@ import time
 
 MODS: list[str] = [
     'clock',
+    'goal',
     'quotes',
     'rotator',
+    'twitchchat',
 ]
+
+
+def fancycb() -> str:
+    cb: str = ''
+
+    for n in str(int(time.time())):
+        cb += chr(int(n) + 97)
+
+    return cb
+
 
 
 if __name__ == '__main__':
@@ -18,11 +30,9 @@ if __name__ == '__main__':
     OUT_DIR: pathlib.Path = REPO_DIR.joinpath('olay', 'mod')
     HEADER_FILE: pathlib.Path = SRC_DIR.joinpath('mod', 'header.html')
     FOOTER_FILE: pathlib.Path = SRC_DIR.joinpath('mod', 'footer.html')
-    CACHE_BUST: str = str(time.time())
+    CACHE_BUST: str = fancycb()
 
     for mod in MODS:
-        print(f'baking {mod} html ...')
-
         body_file: pathlib.Path = MOD_DIR.joinpath(mod, 'body.html')
         out_file: pathlib.Path = OUT_DIR.joinpath(mod, 'index.html')
 
@@ -30,7 +40,10 @@ if __name__ == '__main__':
         code += HEADER_FILE.read_text()
         code += body_file.read_text()
         code += FOOTER_FILE.read_text()
+        code = code.strip()
 
         code = code.replace('{CACHE_BUST}', CACHE_BUST)
 
-        out_file.write_text(code)
+        if not out_file.is_file() or len(code) != len(out_file.read_text().strip()):
+            print(f'{time.time()} baking {mod} html ...')
+            out_file.write_text(code)
